services/telegram-parser/requirements.txt
telethon>=1.36
pyzmq>=25.0
python-dotenv>=1.0
